requirements.txt
langgraph==0.5.0
langchain-groq
langchain-core
python-dotenv
python-docx
